+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## UNRELEASED
 
-* `nil` input is always resolved as `nil`. Fixes [#41](https://github.com/metosin/muuntaja/issues/41)
+* **BREAKING**: Handling empty responses
+  * `:allow-empty-input-on-decode?` is now called `:allow-empty-input`, defaulting to `true`
+  * by default, empty input(streams) is decoded into`nil`
+  * setting the value to `false` with cause the decoder to do whatever it does
+
 * updated deps:
 
 ```clj